crud

+++ Visit.Domain/AttributeValue.cs
@@ -2,13 +2,12 @@
 
 public class AttributeValue
 {
-    
     public long Id { get; set; }
 
     public int PlaceId { get; set; }
     public Place Place { get; set; }
 
-    public long AttributeId { get; set; }
+    public int AttributeId { get; set; }
     public Attribute Attribute { get; set; }
 
     public int? IntValue { get; set; }
